requirements.txt
websocket-client
influxdb
